Refactor: now feature is called live-mount
