+++ @xen-orchestra/mixins/live-mount/_LocalBlockDevice.mjs
